Add 3D gaussian paint dots, with the mouse camera rotating
This is mostly practice for understanding the P3D co-ordinate system,
and understanding how the camera can be panned around a 3D space based
on a mouse position in 2D space (i.e. the location in the canvas.)

This also stresses the point that each frame in processing is new. To
create the illusion of an object in space that you rotate around you
need to redraw it for each frame. You can't move the existing object, or
rotate the entire canvas around. Objects are just pixels, and processing
isn't a physics engine.