Start migration to p5.js with 3D gaussian balls and camera test
Processing sketchbooks are fine, but live sketches on the web are nicer
to "show off", and structuring the sketches in web pages gives me the
option to talk about what the sketch is trying to do or show.

The 3D gaussian balls sketch renders at 50 FPS (capped?) in safari on my macbook,
so performance seems fine so far (but it's terrible in Chrome -- less
that 10 FPS at 500 balls, but then again the github.com/home globe
animation is also super choppy in Chrome.)

I'm keeping the old processing sketches where they are while I play with
this: I may reach the bounds of p5.js quickly (or not at all, and
migrate everything at some point) or even abandon it in favour of
three.js or similar.